Activate only connections if present in the current system
With activate_connections set to true wicked2nm would try to set connections to
the 'up' status even if the interface isn't present in the system resulting in
some cryptic errors by NetworkManager. Now it checks what interfaces are present
and only sets connection with a matching interface present to 'up'.